chore: isolate deploy script.
